Switched to using latest interfacing functions (#360)
* switched to using latest interfacing functions

* Updated get_svc_uri and interface_with_covalent

* Updated changelog
