+++ covalent/__init__.py
@@ -20,13 +20,19 @@
 
 """Main Covalent public functionality."""
 
+
+import contextlib
 import os
 
 from . import executor
-from ._dispatcher_plugins import local_dispatch as dispatch
-from ._dispatcher_plugins import local_dispatch_sync as dispatch_sync
-from ._results_manager.results_manager import cancel, get_result, sync
+from ._results_manager.results_manager import sync
 from ._shared_files.config import get_config, reload_config, set_config
+from ._shared_files.interface_with_covalent import (
+    cancel_workflow,
+    dispatch,
+    dispatch_sync,
+    get_result,
+)
 from ._shared_files.util_classes import RESULT_STATUS as status
 from ._workflow import Lepton
 from ._workflow import electron_func as electron
