+++ refactor/results/main.py
@@ -20,10 +20,10 @@
 
 from pathlib import Path
 
+from app.api.api_v0.api import api_router
+from app.core.config import settings
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from results.app.api.api_v0.api import api_router
-from results.app.core.config import settings
 
 BASE_PATH = Path(__file__).resolve().parent
